# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iris API",
    "description": "Self-hosted media aggregator + streamer — internal household API.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "me"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "The authenticated user's profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/me/continue-watching": {
      "get": {
        "tags": [
          "me"
        ],
        "operationId": "continue_watching",
        "responses": {
          "200": {
            "description": "Recently-watched, not-yet-finished items for resume",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContinueWatchingItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
    "/api/me/display-name": {
      "post": {
        "tags": [
          "me"
        ],
        "operationId": "change_display_name",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeDisplayNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Display name updated"
          },
          "400": {
            "description": "Empty / too-long display name (max 64)"
          }
        }
      }
    },
    "/api/me/follows": {
      "get": {
        "tags": [
          "follows"
        ],
        "operationId": "list_follows",
        "responses": {
          "200": {
            "description": "The caller's followed series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "follows"
        ],
        "operationId": "create_follow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created (or existing) follow summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowSummary"
                }
              }
            }
          },
          "400": {
            "description": "Empty / non-normalisable name"
          }
        }
      }
    },
...
    "/api/me/password": {
      "post": {
        "tags": [
          "me"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed; all other sessions revoked"
          },
          "400": {
            "description": "New password too short (min 8 chars)"
          },
          "401": {
            "description": "Old password mismatch / not authenticated"
          }
        }
      }
    },
...
    "/api/me/watchlist": {
      "get": {
        "tags": [
          "me"
        ],
        "operationId": "list_watchlist",
        "responses": {
          "200": {
            "description": "The caller's per-user Watchlist (TV collections)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeDisplayNameRequest": {
        "type": "object",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string"
          },
          "old_password": {
            "type": "string"
          }
        }
      },
...
      "ContinueWatchingItem": {
        "type": "object",
        "required": [
          "infohash",
          "torrent_name",
          "tmdb_verified",
          "file_idx",
          "position_seconds",
          "last_watched_at",
          "completed"
        ],
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "duration_seconds": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "file_idx": {
            "type": "integer",
            "format": "int64"
          },
          "file_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "infohash": {
            "type": "string"
          },
          "kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MediaKind",
                "description": "`\"movie\"` / `\"tv\"` from the parent collection. Clients pass\nthis to `/api/metadata/tmdb/{id}?kind=` — without it, TMDB's\nseparate id namespaces collide and the lookup serves a\nstranger's poster."
              }
            ]
          },
          "last_watched_at": {
            "type": "string",
            "format": "date-time"
          },
          "position_seconds": {
            "type": "number",
            "format": "double"
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tmdb_verified": {
            "type": "boolean"
          },
          "torrent_name": {
            "type": "string"
          }
        }
      },
...
      "CreateFollowRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The display name from whatever surface the user clicked\n(Discovery / Search / `CollectionPage`). Server normalises it\nfor identity; the original is kept for indexer queries and\nUI display."
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional TMDB id — stored as decoration. Surfaces a poster\nonly after the corresponding collection gets `tmdb_verified`."
          }
        }
      },
...
      "FollowSummary": {
        "type": "object",
        "required": [
          "id",
          "normalized_name",
          "name",
          "new_count",
          "created_at"
        ],
        "properties": {
          "backdrop_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_visited_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "new_count": {
            "type": "integer",
            "format": "int64"
          },
          "normalized_name": {
            "type": "string",
            "description": "SCENE-normalised name — clients route by this, not `tmdb_id`."
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Decoration TMDB id (may be null). Even when present, only\nrendered as a poster after the joined collection is verified."
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "is_admin"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_admin": {
            "type": "boolean"
          }
        }
      },
...
      "MediaKind": {
        "type": "string",
        "enum": [
          "movie",
          "tv"
        ]
      },
...
      "WatchlistItem": {
        "type": "object",
        "description": "Post-0.4 \"My Watchlist\" payload. Derived from TV collections\nthat have at least one ingested episode — the household\nauto-tracks every show they're watching, no Follow button. The\nshape mirrors what the legacy `/api/me/follows` façade returns\nso the web client can flip endpoints without rewriting card\nrendering. Old APK 0.3.1 keeps calling `/api/me/follows`.",
        "required": [
          "id",
          "normalized_name",
          "name",
          "new_count",
          "created_at"
        ],
        "properties": {
          "backdrop_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Collection id — clients route to `/collection/:id`."
          },
          "last_visited_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "new_count": {
            "type": "integer",
            "format": "int64",
            "description": "Distinct (season, episode) the indexer has surfaced since\nthe requesting user last opened this collection. Drives the\n\"X new\" tile badge."
          },
          "normalized_name": {
            "type": "string",
            "description": "SCENE-normalised name. Clients use this to detect \"is this\nsearch result already on my Watchlist?\" without having to\nrun the same normaliser themselves."
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}